Generalize, fix laser/optical being blockable
