+++ src/main/java/gregtech/client/renderer/pipe/PipeRenderer.java
@@ -22,12 +22,15 @@ import gregtech.api.pipenet.block.ItemBlockPipe;
 import gregtech.api.pipenet.block.material.BlockMaterialPipe;
 import gregtech.api.pipenet.block.material.TileEntityMaterialPipeBase;
 import gregtech.api.pipenet.tile.IPipeTile;
+import gregtech.api.pipenet.tile.TileEntityPipeBase;
 import gregtech.api.unification.material.Material;
 import gregtech.api.unification.material.info.MaterialIconType;
 import gregtech.api.util.GTUtility;
 import gregtech.api.util.ModCompatibility;
 import gregtech.client.renderer.CubeRendererState;
 import gregtech.client.renderer.texture.Textures;
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.BufferBuilder;
@@ -58,6 +61,7 @@ import org.lwjgl.opengl.GL11;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
+import java.util.EnumMap;
 import java.util.List;
 
 @SideOnly(Side.CLIENT)
@@ -68,6 +72,36 @@ public abstract class PipeRenderer implements ICCBlockRenderer, IItemRenderer {
     private EnumBlockRenderType blockRenderType;
     protected static final ThreadLocal<BlockRenderer.BlockFace> blockFaces = ThreadLocal.withInitial(BlockRenderer.BlockFace::new);
     private static final Cuboid6 FRAME_RENDER_CUBOID = new Cuboid6(0.001, 0.001, 0.001, 0.999, 0.999, 0.999);
+    private static final EnumMap<EnumFacing, EnumMap<FluidPipeRenderer.Border, EnumFacing>> FACE_BORDER_MAP = new EnumMap<>(EnumFacing.class);
+    private static final Int2ObjectMap<IVertexOperation[]> RESTRICTOR_MAP = new Int2ObjectOpenHashMap<>();
+
+    @SuppressWarnings("unused")
+    public static void initializeRestrictor(TextureMap map) {
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_UP, Border.TOP);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_DOWN, Border.BOTTOM);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_UD, Border.TOP, Border.BOTTOM);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_LEFT, Border.LEFT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_UL, Border.TOP, Border.LEFT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_DL, Border.BOTTOM, Border.LEFT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_NR, Border.TOP, Border.BOTTOM, Border.LEFT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_RIGHT, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_UR, Border.TOP, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_DR, Border.BOTTOM, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_NL, Border.TOP, Border.BOTTOM, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_LR, Border.LEFT, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_ND, Border.TOP, Border.LEFT, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY_NU, Border.BOTTOM, Border.LEFT, Border.RIGHT);
+        addRestrictor(Textures.PIPE_BLOCKED_OVERLAY, Border.TOP, Border.BOTTOM, Border.LEFT, Border.RIGHT);
+    }
+
+    static {
+        FACE_BORDER_MAP.put(EnumFacing.DOWN, borderMap(EnumFacing.NORTH, EnumFacing.SOUTH, EnumFacing.EAST, EnumFacing.WEST));
+        FACE_BORDER_MAP.put(EnumFacing.UP, borderMap(EnumFacing.NORTH, EnumFacing.SOUTH, EnumFacing.WEST, EnumFacing.EAST));
+        FACE_BORDER_MAP.put(EnumFacing.NORTH, borderMap(EnumFacing.UP, EnumFacing.DOWN, EnumFacing.EAST, EnumFacing.WEST));
+        FACE_BORDER_MAP.put(EnumFacing.SOUTH, borderMap(EnumFacing.UP, EnumFacing.DOWN, EnumFacing.WEST, EnumFacing.EAST));
+        FACE_BORDER_MAP.put(EnumFacing.WEST, borderMap(EnumFacing.UP, EnumFacing.DOWN, EnumFacing.NORTH, EnumFacing.SOUTH));
+        FACE_BORDER_MAP.put(EnumFacing.EAST, borderMap(EnumFacing.UP, EnumFacing.DOWN, EnumFacing.SOUTH, EnumFacing.NORTH));
+    }
 
     public PipeRenderer(String name, ModelResourceLocation modelLocation) {
         this.name = name;
@@ -267,6 +301,19 @@ public abstract class PipeRenderer implements ICCBlockRenderer, IItemRenderer {
         for (IVertexOperation[] vertexOperations : renderContext.pipeSideRenderer) {
             renderFace(renderState, vertexOperations, side, cuboid6);
         }
+        int blockedConnections = renderContext.getBlockedConnections();
+        if (blockedConnections != 0) {
+            int borderMask = 0;
+            for (FluidPipeRenderer.Border border : FluidPipeRenderer.Border.VALUES) {
+                if (TileEntityPipeBase.isFaceBlocked(blockedConnections, getSideAtBorder(side, border))) {
+                    borderMask |= border.mask;
+                }
+            }
+            if (borderMask != 0) {
+                IVertexOperation[] pipeline = ArrayUtils.addAll(renderContext.getBaseVertexOperation(), RESTRICTOR_MAP.get(borderMask));
+                renderFace(renderState, pipeline, side, cuboid6);
+            }
+        }
     }
 
     protected void renderFace(CCRenderState renderState, IVertexOperation[] pipeline, EnumFacing side, Cuboid6 cuboid6) {
@@ -459,5 +506,38 @@ public abstract class PipeRenderer implements ICCBlockRenderer, IItemRenderer {
             return pipeThickness;
         }
 
+    }
+
+    private static EnumMap<Border, EnumFacing> borderMap(EnumFacing topSide, EnumFacing bottomSide, EnumFacing leftSide, EnumFacing rightSide) {
+        EnumMap<Border, EnumFacing> sideMap = new EnumMap<>(Border.class);
+        sideMap.put(Border.TOP, topSide);
+        sideMap.put(Border.BOTTOM, bottomSide);
+        sideMap.put(Border.LEFT, leftSide);
+        sideMap.put(Border.RIGHT, rightSide);
+        return sideMap;
+    }
+
+    private static void addRestrictor(TextureAtlasSprite sprite, Border... borders) {
+        int mask = 0;
+        for (Border border : borders) {
+            mask |= border.mask;
+        }
+        RESTRICTOR_MAP.put(mask, new IVertexOperation[]{new IconTransformation(sprite)});
+    }
+
+    protected static EnumFacing getSideAtBorder(EnumFacing side, Border border) {
+        return FACE_BORDER_MAP.get(side).get(border);
+    }
+
+    public enum Border {
+        TOP, BOTTOM, LEFT, RIGHT;
+
+        public static final Border[] VALUES = values();
+
+        public final int mask;
+
+        Border() {
+            mask = 1 << this.ordinal();
+        }
     }
 }
